chore: 사용자 open/close window state move folder
